Loading Points & Comparing
Points are all compared to the first point on spline. When a point is exactly the same as first point, stop
the loading. This is how we will load the spline every time.
However, I have to enhance this algorithm by finding the precise
identical point by computing the distance between the two and look
if the distance is very small (take the smallest distance).